setup do app corrigido e adicionado entry_point teste sample adicionado
close #43
