+++ .github/workflows/conversor_divisor.yml
@@ -23,11 +23,8 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install flake8 pytest codecov
+        pip install codecov
         if [ -f requirements-dev.txt ]; then pip install -r requirements-dev.txt; fi
-    - name: Code formatter with Black
-      run: |
-        black -l 79  -t py37 conversor_divisor/
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
@@ -36,7 +33,7 @@ jobs:
         flake8 . --count --exit-zero --max-complexity=12 --max-line-length=79 --statistics
     - name: Test with pytest
       run: |
-        pytest tests --cov=conversor_divisor
+        pytest tests/ -v --cov=conversor_divisor
     - name: codecov
       run: |
         codecov
