+++ FFmpeg/README.md
@@ -1,0 +1,9 @@
+### Baixe os arquivo do ffmpeg presentes neste [LINK](https://drive.google.com/drive/folders/1h_ySXbsgu-hWg_ZKAWhqQx4mCNzQy3eU?usp=sharing) e extraia os aquivos de acordo com essa arvore de diretórios.
+
+        FFmpeg/
+        ├── bin
+        │   ├── ffmpeg.exe
+        │   ├── ffplay.exe
+        │   └── ffprobe.exe
+        ├── doc
+        ├── presets
