+++ README.md
@@ -2,6 +2,7 @@
 ![Conversor & Divisor](https://github.com/hadtrindade/conversor-e-divisor/workflows/Conversor%20&%20Divisor/badge.svg)
 [![Python 3](https://pyup.io/repos/github/hadtrindade/conversor-e-divisor/python-3-shield.svg)](https://pyup.io/repos/github/hadtrindade/conversor-e-divisor/)
 [![CodeFactor](https://www.codefactor.io/repository/github/hadtrindade/conversor-e-divisor/badge)](https://www.codefactor.io/repository/github/hadtrindade/conversor-e-divisor)
+[![codecov](https://codecov.io/gh/hadtrindade/conversor-e-divisor/branch/master/graph/badge.svg?token=OMF5YWX6I6)](https://codecov.io/gh/hadtrindade/conversor-e-divisor)
 [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
 [![Updates](https://pyup.io/repos/github/hadtrindade/conversor-e-divisor/shield.svg)](https://pyup.io/repos/github/hadtrindade/conversor-e-divisor/)
 
@@ -12,9 +13,16 @@
 
 ![Conversor & Divisor](https://i.imgur.com/whAnUwb.gif)
 ---
-#### Como obter o código-fonte
 
-`git clone https://github.com/hadtrindade/conversor-e-divisor`
+#### Como instalar no windows
+`Baixar o instalador` [Link](https://github.com/hadtrindade/conversor-e-divisor/releases)
+
+
+#### Como instalar no linux
+`git clone https://github.com/hadtrindade/conversor-e-divisor`  
+`cd conversor-e-divisor`  
+`make install ou pip install .`
+
 
 
 `Binários do MP4Box e FFmpeg em `[Aqui](https://drive.google.com/drive/folders/1h_ySXbsgu-hWg_ZKAWhqQx4mCNzQy3eU?usp=sharing)
