+++ pytest.ini
@@ -1,0 +1,2 @@
+[pytest]
+qt_api=pyside2
