+++ requirements-dev.txt
@@ -1,6 +1,8 @@
 flake8
 pytest
 pytest-cov
+pytest-mock
+pytest-qt
 black
 PyQt5==5.15.1
 pyqt5-tools==5.15.1.3
