+++ tests/test_sample.py
@@ -1,2 +1,13 @@
-def test_sample():
-    assert 1 == 1
+from conversor_divisor.app import MainWindow
+from conversor_divisor import __version__
+import pytest
+
+
+@pytest.fixture(scope="module")
+def app():
+    widget = MainWindow()
+    return widget
+
+
+def test_version(app, qtbot):
+    assert app.label_version.text() == __version__
