--- ui_conversor_divisor/__init__.py
@@ -1,2 +1,0 @@
-__version__ = "v1.5.1"
-__author__ = "Ivo H. Trindade Silva"
